requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
vaderSentiment>=3.3.2
lxml>=4.9.0
feedparser>=6.0.0
streamlit>=1.30.0
plotly>=5.18.0
transformers>=4.35.0
torch>=2.0.0
google-genai>=0.1.0
python-dotenv>=1.0.0
schedule>=1.2.1
